GH-94 moved SystemProperties into util package, removed isDemo flag and added org.icepdf.core.annotations.interactive.enabled.
