+++ viewer/viewer-awt/src/main/java/org/icepdf/ri/common/SwingViewBuilder.java
@@ -15,12 +15,11 @@
  */
 package org.icepdf.ri.common;
 
-import org.icepdf.core.SystemProperties;
 import org.icepdf.core.pobjects.Page;
 import org.icepdf.core.pobjects.graphics.images.references.ImageReference;
 import org.icepdf.core.pobjects.graphics.images.references.ImageReferenceFactory;
-import org.icepdf.core.util.Defs;
 import org.icepdf.core.util.Library;
+import org.icepdf.core.util.SystemProperties;
 import org.icepdf.ri.common.utility.annotation.AnnotationPanel;
 import org.icepdf.ri.common.utility.annotation.destinations.DestinationsPanel;
 import org.icepdf.ri.common.utility.annotation.markup.MarkupAnnotationPanel;
@@ -306,12 +305,8 @@ public class SwingViewBuilder {
 
     protected static boolean isMacOs;
 
-    private static boolean isDemo;
-
     static {
         isMacOs = SystemProperties.OS_NAME.contains("OS X");
-        // check for demo system property
-        isDemo = Defs.sysPropertyBoolean("org.icepdf.ri.viewer.demo", false);
     }
 
     /**
@@ -1198,11 +1193,6 @@ public class SwingViewBuilder {
             addToToolBar(toolbar, buildFormsToolBar());
         if (propertiesManager.checkAndStoreBooleanProperty(ViewerPropertiesManager.PROPERTY_SHOW_TOOLBAR_SEARCH))
             addToToolBar(toolbar, buildQuickSearchToolBar());
-
-        // we only add the configurable font engin in the demo version
-        if (isDemo) {
-            addToToolBar(toolbar, buildDemoToolBar());
-        }
 
         // Set the toolbar back to null if no components were added
         // The result of this will properly disable the necessary menu items for controlling the toolbar
