+++ viewer/viewer-awt/src/main/java/org/icepdf/ri/util/BareBonesBrowserLaunch.java
@@ -15,7 +15,7 @@
  */
 package org.icepdf.ri.util;
 
-import org.icepdf.core.SystemProperties;
+import org.icepdf.core.util.SystemProperties;
 
 import javax.swing.*;
 import java.lang.reflect.Method;
